Update button label for navigating to next instance variable
When landing on the instance variable page from the search page,
there is a button used to navigate to the next variable in the dataset.
The user might be mistaken to think the button leads to the next hit in
the search results which is not the case. To clarify the situation,
the label for the button has been updated to indicate the association
between instance variable and datasets.

- add the word 'dataset' to move to next -button navigation label in
  editor and catalog

Issue: THLDTKK-558, THLDTKK-566
Change-Id: I3e88755641a906865d3583d415dcdafebcd73f96