Hide button for editing study and dataset persons from basic users
Basic users are not able to modify information about existing persons in
studies and datasets. Despite the fact that storing of edited persons
was disabled on the server side, the client side edit button was still
visible to normal users without administrative privileges.

Users could click on the button and try to edit the details of a person
but saving the person resulted in an exception. Since saving is not
possible for basic users, the edit button has also been removed from the
user interface and made visible only to admins.

- hide edit person button for basic users from editing of studies and
  datasets in editor

Issue: THLDTKK-561
Change-Id: I2c9bd6c875931383830c48ab651760d68544828b