Move editing of person details to maintenance menu
Previously, editing of people had to be done through the edit views
of studies or datasets, and there was not a single place to display
these commonly shared entities in editor. With this update, editing
of people has been moved to its own admin-specific maintenance view
accessed through the maintenance menu. Editing and deleting of people
is only possible for admins

- new maintenance admin menu item for managing details of people used
  in studies and datasets.

Issue: THLDTKK-563, THLDTKK-576
Change-Id: I83c7214d3648e54aafb296ef714fe971b2f1123f