Add 'profiling and automatic decision making' to study editor
Implements option for showing whether study is used for profiling and
automatic decision making. The field will either display additional
details described by the user if the answer is 'yes', or the field will
dynamically tell that the specific organization does not use the data
for profiling and decision making.

Issue: THLDTKK-652
Change-Id: Ica48117a3dd361831136045fe42c396c8c82c8f2