Add principal investigator to scientific privacy notice
The principal investigator (PI) of a study is a person chiefly responsible
for carrying out a scientific endeavour. A research group can also be
charged with the responsibility. In this update, information about the
PI or group is added to the printed scientific privacy notice in its own
section.

- add principal investigator or research group info to the printed
  scientific privacy notice
- info includes name, and phone and email for contact

Issue: THLDTKK-609
Change-Id: Ibc6d18b212ae44cd612e66df82f82df9e1e1a15a