THLDTKK-801: Remove catalog dataset instance variable search toolbar
Although datasets can have hundreds of instance variables, the search
toolbar in the catalog's dataset view had to be removed for making
the user interface more simple. The page navigation has been left
intact but with this update there is no longer the possibility to search
for individual instance variables in a dataset. Minor style changes
were also done to the page navigation toolbar in the same view.

- remove the search toolbar from the catalog's dataset view that allowed
  users to find variables with arbitrary user input
- change page navigation style colors to match theme for instance
  variable navigation in catalog's dataset view

Issue: THLDTKK-801
Change-Id: I2a7899c94aa9c786faef135a9282b7bc976fe7aa