THLDTKK-711: Fix publishing of studies with associated organizations
In editor and in the administrative information of a study, it is
possible to define other organizations which are jointly responsible for
a study. Publishing or actually republishing a study with such
associated organizations to catalog did not work. This happened because
the update and publishing of public studies did not consider the
associated organization references. The same was true for withdrawing
a study from the public catalog, and associated organizations would not
have been deleted.

Publishing and withdrawing of studies with associated
organizations has now been fixed. Also, at the same time publishing and
withdrawing of studies with study forms was fixed so that the updates
consider new study forms and the references get cleaned up from the
public catalog after withdrawal.

- fix catalog study publish updates to consider associated organizations
- fix catalog study publish updates to consider study forms
- fix catalog study withdraw to delete lingering study form references

Issue: THLDTKK-711
Change-Id: I4a334f105ebaf88a9b3d9b827f9d9a601335d398